# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/ib/applications": {
      "get": {
        "description": "Paginated, newest first, with per-status counts for the tab labels. Both the rows and the counts respect the reviewing admin’s client scope.",
        "operationId": "AdminIbController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "approved",
                "rejected"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "submittedAt",
                "status",
                "userEmail",
                "userFirstName"
              ],
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "The partner application queue",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/ib/applications/export": {
      "get": {
        "description": "The same `status` filter as GET /admin/ib/applications, over every matching row rather than one page. Respects the reviewing admin’s client scope.",
        "operationId": "AdminIbController_exportApplications",
        "parameters": [
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "csv"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "approved",
                "rejected"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A CSV file, named `ib-applications-<YYYY-MM-DD>.csv`.",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Export the partner application queue as CSV",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/ib/applications/{id}/approve": {
      "patch": {
        "description": "One transaction: the application moves out of pending and the account is created together, so there is no state where a client has been told they were accepted and has no referral code. Refuses if another reviewer already decided it.",
        "operationId": "AdminIbController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveIbApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IbAccountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Approve an application and create the partner account",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/ib/applications/{id}/reject": {
      "patch": {
        "description": "The reason is required and composed server-side from a configured label plus an optional note. It is stored already composed, because that is the sentence the client is shown.",
        "operationId": "AdminIbController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectIbApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IbApplicationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Reject an application, with a reason",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/ib/partners": {
      "get": {
        "description": "Joined to the person and their level, newest approval first.",
        "operationId": "AdminIbController_listPartners",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "approvedAt",
                "level",
                "referralCode",
                "userEmail",
                "userFirstName"
              ],
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "The partner list",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/ib/partners/export": {
      "get": {
        "description": "Every partner the acting admin may see, joined to the person and their level. The list takes no filters, so neither does its export.",
        "operationId": "AdminIbController_exportPartners",
        "parameters": [
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "csv"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A CSV file, named `ib-partners-<YYYY-MM-DD>.csv`.",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Export the partner list as CSV",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/ib/partners/{userId}/level": {
      "patch": {
        "description": "The target level must be ENABLED — a disabled one takes no share, so placing somebody on it stops their earnings silently rather than demoting them visibly.",
        "operationId": "AdminIbController_changeLevel",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeIbLevelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IbAccountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Move a partner to a different level",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/ib/partners/{userId}/parent": {
      "patch": {
        "description": "Refuses a change that would put a partner beneath their own descendant. A self-FK cannot catch that — Postgres accepts A→B→A — and the payout walk climbs parents until it runs out, so a loop is a walk that never does.",
        "operationId": "AdminIbController_reassignParent",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReassignIbParentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IbAccountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Reassign a partner’s parent",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/ib/partners/{userId}/active": {
      "patch": {
        "description": "Suspension keeps the referral code and the tree and stops the earning. There is no delete: removing the row would orphan every client and partner attributed beneath them.",
        "operationId": "AdminIbController_setActive",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetIbActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IbAccountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Suspend or reactivate a partner",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "OxShare CRM API",
    "description": "Forex/CFD Introducing-Broker CRM — Phase 1",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "oxshare_crm_portal_at",
        "description": "Portal session. Set by POST /v1/auth/login; httpOnly, so not settable here."
      },
      "admin": {
        "type": "apiKey",
        "in": "cookie",
        "name": "oxshare_crm_admin_at",
        "description": "Admin session. Set by POST /v1/admin/auth/login; httpOnly, so not settable here."
      }
    },
    "schemas": {
...
      "IbAccountDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "level": {
            "type": "number"
          },
          "parentIbUserId": {
            "type": "string",
            "nullable": true
          },
          "referralCode": {
            "type": "string",
            "description": "What a client types at registration to be attributed here."
          },
          "active": {
            "type": "boolean"
          },
          "approvedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "level",
          "parentIbUserId",
          "referralCode",
          "active",
          "approvedAt"
        ]
      },
      "IbApplicationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "motivation": {
            "type": "string",
            "nullable": true
          },
          "expectedVolume": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true,
            "description": "Already composed — this is the sentence the client is shown."
          },
          "reviewedBy": {
            "type": "string",
            "nullable": true
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "motivation",
          "expectedVolume",
          "website",
          "status",
          "rejectionReason",
          "reviewedBy",
          "reviewedAt",
          "submittedAt"
        ]
      },
...
      "ApproveIbApplicationDto": {
        "type": "object",
        "properties": {
          "level": {
            "type": "number",
            "minimum": 1,
            "description": "Which rung to place them on. Omitted, the service derives it: the shallowest enabled level with no parent, one below the parent otherwise."
          },
          "parentIbUserId": {
            "type": "string",
            "nullable": true,
            "description": "The partner who introduced them. Omitted or null means they deal direct."
          }
        }
      },
      "RejectIbApplicationDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "A configured label from the `partner` rejection context."
          },
          "note": {
            "type": "string",
            "description": "The reviewer's own words, appended to the label."
          }
        }
      },
      "ChangeIbLevelDto": {
        "type": "object",
        "properties": {
          "level": {
            "type": "number",
            "minimum": 1,
            "description": "Must be an ENABLED level."
          }
        },
        "required": [
          "level"
        ]
      },
      "ReassignIbParentDto": {
        "type": "object",
        "properties": {
          "parentIbUserId": {
            "type": "string",
            "nullable": true,
            "description": "The new parent partner, or null to make them a direct partner."
          }
        },
        "required": [
          "parentIbUserId"
        ]
      },
      "SetIbActiveDto": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "description": "False suspends: the referral code and the tree are kept, the earning stops. There is no delete — removing the row would orphan every partner beneath them."
          }
        },
        "required": [
          "active"
        ]
      },
...
    }
  }
}