# src/modules/ib/admin-ib.controller.ts
import {
  Body,
  Controller,
  Get,
  Param,
  ParseUUIDPipe,
  Patch,
  Query,
  Req,
  Res,
  UseGuards,
} from '@nestjs/common';
import { ApiCookieAuth, ApiOkResponse, ApiOperation, ApiQuery, ApiTags } from '@nestjs/swagger';
import { Request, Response } from 'express';
import {
  AuthenticatedAdmin,
  PermissionsGuard,
  RequirePermissions,
} from '../admin/guards/admin.guard';
import { ScopedToClients } from '../admin/guards/client-scope.decorator';
import { Audited } from '../admin/guards/audited.decorator';
import { IbApplicationsService } from './ib-applications.service';
import { IB_APPLICATION_SORT_COLUMNS, IB_PARTNER_SORT_COLUMNS } from '../../store/ib.store';
import { AdminExportService } from '../admin/admin-export.service';
import { AdminAuditService } from '../admin/admin-audit.service';
import { exportFormat, streamCsv } from '../../common/export/export-response';
import {
  ApproveIbApplicationDto,
  ChangeIbLevelDto,
  IbAccountDto,
  IbApplicationDto,
  IB_APPLICATION_STATUSES,
  ReassignIbParentDto,
  RejectIbApplicationDto,
  SetIbActiveDto,
  type IbApplicationStatusDto,
} from './dto/ib-application.dto';

/**
 * Reviewing partner applications.
 *
 * ## `@ScopedToClients`, unlike the levels controller
 *
 * Every route here reaches a CLIENT — an application names the person who made
 * it — so an admin restricted to a subset of the client base must not see or
 * decide one from outside it. The list applies the predicate in its query; the
 * decisions ask `ClientVisibilityService` first, and get a 404 rather than a
 * 403 so the id cannot be used to enumerate the clients they were denied.
 *
 * ## Approve and reject are SEPARATE permissions
 *
 * `ib.approve` and `ib.reject`, not one `ib.decide`. Granting somebody the
 * ability to turn applications down is a smaller act of trust than granting
 * them the ability to create partners who will be paid, and a single key makes
 * that distinction unavailable to whoever configures a role.
 *
 * Neither one implies `ib.manage`: rewriting the payout ladder is a different
 * power from deciding who joins it.
 */
@ApiTags('admin')
@Controller('admin/ib')
export class AdminIbController {
  constructor(
    private readonly applications: IbApplicationsService,
    private readonly exports: AdminExportService,
    private readonly audit: AdminAuditService,
  ) {}

  @Get('applications')
  @UseGuards(PermissionsGuard)
  @RequirePermissions('ib.view')
  @ApiCookieAuth()
  @ApiOperation({
    summary: 'The partner application queue',
    description:
      'Paginated, newest first, with per-status counts for the tab labels. Both the rows and ' +
      'the counts respect the reviewing admin’s client scope.',
  })
  /*
   * Declared OPTIONAL, explicitly — otherwise Swagger emits every `@Query()` as
   * `required: true` and the generated frontend types demand all five on a call
   * that legitimately passes none.
   */
  @ApiQuery({ name: 'status', required: false, enum: IB_APPLICATION_STATUSES })
  @ApiQuery({ name: 'page', required: false })
  @ApiQuery({ name: 'limit', required: false })
  @ApiQuery({ name: 'sort', required: false, enum: Object.keys(IB_APPLICATION_SORT_COLUMNS) })
  @ApiQuery({ name: 'order', required: false, enum: ['asc', 'desc'] })
  @ScopedToClients('IbStore.findPageWithUsers applies the predicate to ib_applications.user_id.')
  list(
    @Req() req: Request & { admin: AuthenticatedAdmin },
    @Query('status') status?: string,
    @Query('page') page?: string,
    @Query('limit') limit?: string,
    @Query('sort') sort?: string,
    @Query('order') order?: string,
  ) {
    return this.applications.list(
      {
        // `ib_application_status` is a Postgres enum, so an unrecognised value
        // would error in the database rather than at the edge.
        status: parseStatus(status),
        page: parsePositive(page),
        limit: parsePositive(limit),
        // Validated in the service against the allowlist, which is where the
        // column mapping lives.
        sort,
        order,
      },
      req.admin.clientScope,
    );
  }

  /**
   * The partner application queue as CSV.
   *
   * Declared before `applications/:id/*` so Express does not bind `id =
   * 'export'` — the same ordering rule the client and KYC exports follow.
   *
   * `ib.view`, matching the list. Notably NOT `ib.approve`: reading the queue
   * and deciding on it are separate powers here, and an export is a read.
   */
  @Get('applications/export')
  @UseGuards(PermissionsGuard)
  @RequirePermissions('ib.view')
  @ApiCookieAuth()
  @ApiOperation({
    summary: 'Export the partner application queue as CSV',
    description:
      'The same `status` filter as GET /admin/ib/applications, over every matching row rather ' +
      'than one page. Respects the reviewing admin’s client scope.',
  })
  @ApiOkResponse({
    description: 'A CSV file, named `ib-applications-<YYYY-MM-DD>.csv`.',
    content: { 'text/csv': { schema: { type: 'string', format: 'binary' } } },
  })
  @ApiQuery({ name: 'format', required: false, enum: ['csv'] })
  @ApiQuery({ name: 'status', required: false, enum: IB_APPLICATION_STATUSES })
  @ScopedToClients(
    'AdminExportService.ibApplicationBatch → IbStore.findPageWithUsers with actor.clientScope, the same predicate on users.id the queue applies.',
  )
  @Audited('export.ib_applications')
  async exportApplications(
    @Req() req: Request & { admin: AuthenticatedAdmin },
    @Res() res: Response,
    @Query('format') format?: string,
    @Query('status') status?: string,
  ) {
    const chosen = exportFormat(format);
    const query = { status: parseStatus(status) };

    this.audit.record(req.admin.id, 'export.ib_applications', 'ib_applications', req.admin.id, {
      format: chosen,
      filters: query,
    });

    await streamCsv(
      res,
      'ib-applications',
      chosen,
      this.exports.ibApplicationColumns,
      (offset, limit) => this.exports.ibApplicationBatch(query, req.admin, offset, limit),
    );
  }

  @Patch('applications/:id/approve')
  @UseGuards(PermissionsGuard)
  @RequirePermissions('ib.approve')
  @ApiCookieAuth()
  @ApiOperation({
    summary: 'Approve an application and create the partner account',
    description:
      'One transaction: the application moves out of pending and the account is created ' +
      'together, so there is no state where a client has been told they were accepted and has ' +
      'no referral code. Refuses if another reviewer already decided it.',
  })
  @ApiOkResponse({ type: IbAccountDto })
  @ScopedToClients('Decides on one client’s application; out-of-scope 404s like a missing one.')
  @Audited('ib.approve')
  approve(
    @Req() req: Request & { admin: AuthenticatedAdmin },
    @Param('id', ParseUUIDPipe) id: string,
    @Body() dto: ApproveIbApplicationDto,
  ) {
    return this.applications.approve(id, req.admin.id, req.admin.clientScope, {
      level: dto.level,
      parentIbUserId: dto.parentIbUserId ?? null,
    });
  }

  @Patch('applications/:id/reject')
  @UseGuards(PermissionsGuard)
  @RequirePermissions('ib.reject')
  @ApiCookieAuth()
  @ApiOperation({
    summary: 'Reject an application, with a reason',
    description:
      'The reason is required and composed server-side from a configured label plus an optional ' +
      'note. It is stored already composed, because that is the sentence the client is shown.',
  })
  @ApiOkResponse({ type: IbApplicationDto })
  @ScopedToClients('Decides on one client’s application; out-of-scope 404s like a missing one.')
  @Audited('ib.reject')
  reject(
    @Req() req: Request & { admin: AuthenticatedAdmin },
    @Param('id', ParseUUIDPipe) id: string,
    @Body() dto: RejectIbApplicationDto,
  ) {
    return this.applications.reject(id, req.admin.id, req.admin.clientScope, dto);
  }

  // ── partners, once they exist ──────────────────────────────────────────────

  @Get('partners')
  @UseGuards(PermissionsGuard)
  @RequirePermissions('ib.view')
  @ApiCookieAuth()
  @ApiOperation({
    summary: 'The partner list',
    description: 'Joined to the person and their level, newest approval first.',
  })
  @ApiQuery({ name: 'page', required: false })
  @ApiQuery({ name: 'limit', required: false })
  @ApiQuery({ name: 'sort', required: false, enum: Object.keys(IB_PARTNER_SORT_COLUMNS) })
  @ApiQuery({ name: 'order', required: false, enum: ['asc', 'desc'] })
  @ScopedToClients('IbStore.findPartnersPage applies the predicate to ib_accounts.user_id.')
  listPartners(
    @Req() req: Request & { admin: AuthenticatedAdmin },
    @Query('page') page?: string,
    @Query('limit') limit?: string,
    @Query('sort') sort?: string,
    @Query('order') order?: string,
  ) {
    return this.applications.listPartners(
      { page: parsePositive(page), limit: parsePositive(limit), sort, order },
      req.admin.clientScope,
    );
  }

  /**
   * The partner list as CSV.
   *
   * Declared before `partners/:userId/*` for the same routing reason as the
   * applications export above.
   */
  @Get('partners/export')
  @UseGuards(PermissionsGuard)
  @RequirePermissions('ib.view')
  @ApiCookieAuth()
  @ApiOperation({
    summary: 'Export the partner list as CSV',
    description:
      'Every partner the acting admin may see, joined to the person and their level. The list ' +
      'takes no filters, so neither does its export.',
  })
  @ApiOkResponse({
    description: 'A CSV file, named `ib-partners-<YYYY-MM-DD>.csv`.',
    content: { 'text/csv': { schema: { type: 'string', format: 'binary' } } },
  })
  @ApiQuery({ name: 'format', required: false, enum: ['csv'] })
  @ScopedToClients(
    'AdminExportService.ibPartnerBatch → IbStore.findPartnersPage with actor.clientScope, the same predicate on users.id the list applies.',
  )
  @Audited('export.ib_partners')
  async exportPartners(
    @Req() req: Request & { admin: AuthenticatedAdmin },
    @Res() res: Response,
    @Query('format') format?: string,
  ) {
    const chosen = exportFormat(format);

    this.audit.record(req.admin.id, 'export.ib_partners', 'ib_partners', req.admin.id, {
      format: chosen,
    });

    await streamCsv(res, 'ib-partners', chosen, this.exports.ibPartnerColumns, (offset, limit) =>
      this.exports.ibPartnerBatch(req.admin, offset, limit),
    );
  }

  @Patch('partners/:userId/level')
  @UseGuards(PermissionsGuard)
  @RequirePermissions('ib.manage')
  @ApiCookieAuth()
  @ApiOperation({
    summary: 'Move a partner to a different level',
    description:
      'The target level must be ENABLED — a disabled one takes no share, so placing somebody on ' +
      'it stops their earnings silently rather than demoting them visibly.',
  })
  @ApiOkResponse({ type: IbAccountDto })
  @ScopedToClients('Acts on one client’s partner account; out-of-scope 404s like a missing one.')
  @Audited('ib.level_change')
  changeLevel(
    @Req() req: Request & { admin: AuthenticatedAdmin },
    @Param('userId', ParseUUIDPipe) userId: string,
    @Body() dto: ChangeIbLevelDto,
  ) {
    return this.applications.changeLevel(userId, dto.level, req.admin.clientScope);
  }

  @Patch('partners/:userId/parent')
  @UseGuards(PermissionsGuard)
  @RequirePermissions('ib.manage')
  @ApiCookieAuth()
  @ApiOperation({
    summary: 'Reassign a partner’s parent',
    description:
      'Refuses a change that would put a partner beneath their own descendant. A self-FK cannot ' +
      'catch that — Postgres accepts A→B→A — and the payout walk climbs parents until it runs ' +
      'out, so a loop is a walk that never does.',
  })
  @ApiOkResponse({ type: IbAccountDto })
  @ScopedToClients('Acts on one client’s partner account; out-of-scope 404s like a missing one.')
  @Audited('ib.parent_change')
  reassignParent(
    @Req() req: Request & { admin: AuthenticatedAdmin },
    @Param('userId', ParseUUIDPipe) userId: string,
    @Body() dto: ReassignIbParentDto,
  ) {
    return this.applications.reassignParent(userId, dto.parentIbUserId, req.admin.clientScope);
  }

  @Patch('partners/:userId/active')
  @UseGuards(PermissionsGuard)
  @RequirePermissions('ib.manage')
  @ApiCookieAuth()
  @ApiOperation({
    summary: 'Suspend or reactivate a partner',
    description:
      'Suspension keeps the referral code and the tree and stops the earning. There is no ' +
      'delete: removing the row would orphan every client and partner attributed beneath them.',
  })
  @ApiOkResponse({ type: IbAccountDto })
  @ScopedToClients('Acts on one client’s partner account; out-of-scope 404s like a missing one.')
  @Audited('ib.suspend')
  setActive(
    @Req() req: Request & { admin: AuthenticatedAdmin },
    @Param('userId', ParseUUIDPipe) userId: string,
    @Body() dto: SetIbActiveDto,
  ) {
    return this.applications.setActive(userId, dto.active, req.admin.clientScope);
  }
}

/** An unrecognised status is ignored rather than 500ing in the database. */
function parseStatus(value?: string): IbApplicationStatusDto | undefined {
  return IB_APPLICATION_STATUSES.find((s) => s === value);
}

function parsePositive(value?: string): number | undefined {
  const parsed = Number(value);
  return Number.isInteger(parsed) && parsed > 0 ? parsed : undefined;
}